Move int to hex encoding for cids into AtLib
Before, this happened in some places in the GS_API part, and left it to
the user for a lot of other AtLib functions. Since this encoding is
really a detail of the AT command encoding, doing the encoding while
formatting the command seems like the proper place.

Now, almost all cids passed around are just plain integers instead of
ASCII-encoded hex characters as before.